% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkItEngine.R
\name{LinkItEngine}
\alias{LinkItEngine}
\title{Business logic of the LinkIt application}
\description{
An R6 object with methods for use in the Shiny application `ede_linkt_app`, 
and `ede-izm_frontend`.
}
\section{Super class}{
\code{shintodb::NA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{LinkItEngine$new()}}
\item \href{#method-NA-store_user_object}{\code{LinkItEngine$store_user_object()}}
\item \href{#method-NA-save_coordinatoren_list}{\code{LinkItEngine$save_coordinatoren_list()}}
\item \href{#method-NA-username_from_userid}{\code{LinkItEngine$username_from_userid()}}
\item \href{#method-NA-store_msal_user_object}{\code{LinkItEngine$store_msal_user_object()}}
\item \href{#method-NA-save_msal_coordinatoren_list}{\code{LinkItEngine$save_msal_coordinatoren_list()}}
\item \href{#method-NA-username_from_msal_userid}{\code{LinkItEngine$username_from_msal_userid()}}
\item \href{#method-NA-delete_rows_where}{\code{LinkItEngine$delete_rows_where()}}
\item \href{#method-NA-delete_rows_where2}{\code{LinkItEngine$delete_rows_where2()}}
\item \href{#method-NA-make_choices}{\code{LinkItEngine$make_choices()}}
\item \href{#method-NA-get_favorites}{\code{LinkItEngine$get_favorites()}}
\item \href{#method-NA-is_favorite}{\code{LinkItEngine$is_favorite()}}
\item \href{#method-NA-format_dossierid}{\code{LinkItEngine$format_dossierid()}}
\item \href{#method-NA-add_favorite}{\code{LinkItEngine$add_favorite()}}
\item \href{#method-NA-delete_favorite}{\code{LinkItEngine$delete_favorite()}}
\item \href{#method-NA-toggle_favorite}{\code{LinkItEngine$toggle_favorite()}}
\item \href{#method-NA-next_dossier_id}{\code{LinkItEngine$next_dossier_id()}}
\item \href{#method-NA-add_dossier}{\code{LinkItEngine$add_dossier()}}
\item \href{#method-NA-get_dossier}{\code{LinkItEngine$get_dossier()}}
\item \href{#method-NA-get_my_active_dossiers}{\code{LinkItEngine$get_my_active_dossiers()}}
\item \href{#method-NA-get_all_active_dossiers}{\code{LinkItEngine$get_all_active_dossiers()}}
\item \href{#method-NA-get_verwijderde_dossiers}{\code{LinkItEngine$get_verwijderde_dossiers()}}
\item \href{#method-NA-get_personen_active_dossiers}{\code{LinkItEngine$get_personen_active_dossiers()}}
\item \href{#method-NA-get_objecten_objectinstance}{\code{LinkItEngine$get_objecten_objectinstance()}}
\item \href{#method-NA-get_dossierid_personen}{\code{LinkItEngine$get_dossierid_personen()}}
\item \href{#method-NA-dossier_naam_from_id}{\code{LinkItEngine$dossier_naam_from_id()}}
\item \href{#method-NA-delete_dossier}{\code{LinkItEngine$delete_dossier()}}
\item \href{#method-NA-verwijder_dossier}{\code{LinkItEngine$verwijder_dossier()}}
\item \href{#method-NA-undo_verwijder_dossier}{\code{LinkItEngine$undo_verwijder_dossier()}}
\item \href{#method-NA-dossier_exists}{\code{LinkItEngine$dossier_exists()}}
\item \href{#method-NA-add_instantie}{\code{LinkItEngine$add_instantie()}}
\item \href{#method-NA-list_instanties}{\code{LinkItEngine$list_instanties()}}
\item \href{#method-NA-log_audit}{\code{LinkItEngine$log_audit()}}
\item \href{#method-NA-get_audit}{\code{LinkItEngine$get_audit()}}
\item \href{#method-NA-is_object_in_dossier}{\code{LinkItEngine$is_object_in_dossier()}}
\item \href{#method-NA-add_object_to_dossier}{\code{LinkItEngine$add_object_to_dossier()}}
\item \href{#method-NA-remove_object_from_dossier}{\code{LinkItEngine$remove_object_from_dossier()}}
\item \href{#method-NA-add_persoon_kenmerken}{\code{LinkItEngine$add_persoon_kenmerken()}}
\item \href{#method-NA-add_unknown_person}{\code{LinkItEngine$add_unknown_person()}}
\item \href{#method-NA-get_persoon}{\code{LinkItEngine$get_persoon()}}
\item \href{#method-NA-edit_unknown_persoon}{\code{LinkItEngine$edit_unknown_persoon()}}
\item \href{#method-NA-add_unknown_location}{\code{LinkItEngine$add_unknown_location()}}
\item \href{#method-NA-delete_unknown_location}{\code{LinkItEngine$delete_unknown_location()}}
\item \href{#method-NA-get_unknown_location}{\code{LinkItEngine$get_unknown_location()}}
\item \href{#method-NA-set_adres_as_hoofdadres}{\code{LinkItEngine$set_adres_as_hoofdadres()}}
\item \href{#method-NA-set_object_relations}{\code{LinkItEngine$set_object_relations()}}
\item \href{#method-NA-get_object_kenmerk}{\code{LinkItEngine$get_object_kenmerk()}}
\item \href{#method-NA-set_object_kenmerk}{\code{LinkItEngine$set_object_kenmerk()}}
\item \href{#method-NA-get_objects}{\code{LinkItEngine$get_objects()}}
\item \href{#method-NA-n_personen_dossier}{\code{LinkItEngine$n_personen_dossier()}}
\item \href{#method-NA-add_dossier_kenmerken}{\code{LinkItEngine$add_dossier_kenmerken()}}
\item \href{#method-NA-get_dossier_kenmerken}{\code{LinkItEngine$get_dossier_kenmerken()}}
\item \href{#method-NA-replace_dossier_kenmerken}{\code{LinkItEngine$replace_dossier_kenmerken()}}
\item \href{#method-NA-get_dossierid_dossier_kenmerken}{\code{LinkItEngine$get_dossierid_dossier_kenmerken()}}
\item \href{#method-NA-add_coordinatoren}{\code{LinkItEngine$add_coordinatoren()}}
\item \href{#method-NA-delete_dossier_coordinatoren}{\code{LinkItEngine$delete_dossier_coordinatoren()}}
\item \href{#method-NA-replace_dossier_coordinatoren}{\code{LinkItEngine$replace_dossier_coordinatoren()}}
\item \href{#method-NA-get_coordinatoren}{\code{LinkItEngine$get_coordinatoren()}}
\item \href{#method-NA-get_dossierid_coordinatoren}{\code{LinkItEngine$get_dossierid_coordinatoren()}}
\item \href{#method-NA-encrypt}{\code{LinkItEngine$encrypt()}}
\item \href{#method-NA-decrypt}{\code{LinkItEngine$decrypt()}}
\item \href{#method-unknown-clone}{\code{LinkItEngine$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="append_data"><a href='../../shintodb/html/NA.html#method-NA-append_data'><code>shintodb::NA$append_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="close"><a href='../../shintodb/html/NA.html#method-NA-close'><code>shintodb::NA$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="connect_to_database"><a href='../../shintodb/html/NA.html#method-NA-connect_to_database'><code>shintodb::NA$connect_to_database()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="execute_query"><a href='../../shintodb/html/NA.html#method-NA-execute_query'><code>shintodb::NA$execute_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="filter"><a href='../../shintodb/html/NA.html#method-NA-filter'><code>shintodb::NA$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="get_query"><a href='../../shintodb/html/NA.html#method-NA-get_query'><code>shintodb::NA$get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="has_connection"><a href='../../shintodb/html/NA.html#method-NA-has_connection'><code>shintodb::NA$has_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="has_value"><a href='../../shintodb/html/NA.html#method-NA-has_value'><code>shintodb::NA$has_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="have_column"><a href='../../shintodb/html/NA.html#method-NA-have_column'><code>shintodb::NA$have_column()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="list_indexes"><a href='../../shintodb/html/NA.html#method-NA-list_indexes'><code>shintodb::NA$list_indexes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="list_tables"><a href='../../shintodb/html/NA.html#method-NA-list_tables'><code>shintodb::NA$list_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="make_column"><a href='../../shintodb/html/NA.html#method-NA-make_column'><code>shintodb::NA$make_column()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="postgres_now"><a href='../../shintodb/html/NA.html#method-NA-postgres_now'><code>shintodb::NA$postgres_now()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="postgres_now_in_local_time"><a href='../../shintodb/html/NA.html#method-NA-postgres_now_in_local_time'><code>shintodb::NA$postgres_now_in_local_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="query"><a href='../../shintodb/html/NA.html#method-NA-query'><code>shintodb::NA$query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="read_table"><a href='../../shintodb/html/NA.html#method-NA-read_table'><code>shintodb::NA$read_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="replace_value_where"><a href='../../shintodb/html/NA.html#method-NA-replace_value_where'><code>shintodb::NA$replace_value_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="replace_value_where2"><a href='../../shintodb/html/NA.html#method-NA-replace_value_where2'><code>shintodb::NA$replace_value_where2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="table_columns"><a href='../../shintodb/html/NA.html#method-NA-table_columns'><code>shintodb::NA$table_columns()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="table_info"><a href='../../shintodb/html/NA.html#method-NA-table_info'><code>shintodb::NA$table_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="update_where"><a href='../../shintodb/html/NA.html#method-NA-update_where'><code>shintodb::NA$update_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="write_table_overwrite"><a href='../../shintodb/html/NA.html#method-NA-write_table_overwrite'><code>shintodb::NA$write_table_overwrite()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$new(
  gemeente = NULL,
  schema,
  pool,
  config_file = "conf/config.yml",
  what = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-store_user_object"></a>}}
\if{latex}{\out{\hypertarget{method-NA-store_user_object}{}}}
\subsection{Method \code{store_user_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$store_user_object(.user)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-save_coordinatoren_list"></a>}}
\if{latex}{\out{\hypertarget{method-NA-save_coordinatoren_list}{}}}
\subsection{Method \code{save_coordinatoren_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$save_coordinatoren_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-username_from_userid"></a>}}
\if{latex}{\out{\hypertarget{method-NA-username_from_userid}{}}}
\subsection{Method \code{username_from_userid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$username_from_userid(userid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-store_msal_user_object"></a>}}
\if{latex}{\out{\hypertarget{method-NA-store_msal_user_object}{}}}
\subsection{Method \code{store_msal_user_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$store_msal_user_object(.user)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-save_msal_coordinatoren_list"></a>}}
\if{latex}{\out{\hypertarget{method-NA-save_msal_coordinatoren_list}{}}}
\subsection{Method \code{save_msal_coordinatoren_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$save_msal_coordinatoren_list(appname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-username_from_msal_userid"></a>}}
\if{latex}{\out{\hypertarget{method-NA-username_from_msal_userid}{}}}
\subsection{Method \code{username_from_msal_userid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$username_from_msal_userid(userid, appname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-delete_rows_where"></a>}}
\if{latex}{\out{\hypertarget{method-NA-delete_rows_where}{}}}
\subsection{Method \code{delete_rows_where()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$delete_rows_where(table, col_compare, val_compare)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-delete_rows_where2"></a>}}
\if{latex}{\out{\hypertarget{method-NA-delete_rows_where2}{}}}
\subsection{Method \code{delete_rows_where2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$delete_rows_where2(
  table,
  col_compare,
  val_compare,
  col_compare2,
  val_compare2
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-make_choices"></a>}}
\if{latex}{\out{\hypertarget{method-NA-make_choices}{}}}
\subsection{Method \code{make_choices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$make_choices(values_from, names_from = values_from, data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_favorites"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_favorites}{}}}
\subsection{Method \code{get_favorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_favorites(userid, lazy = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-is_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-NA-is_favorite}{}}}
\subsection{Method \code{is_favorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$is_favorite(userid, dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-format_dossierid"></a>}}
\if{latex}{\out{\hypertarget{method-NA-format_dossierid}{}}}
\subsection{Method \code{format_dossierid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$format_dossierid(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_favorite}{}}}
\subsection{Method \code{add_favorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$add_favorite(userid, dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-delete_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-NA-delete_favorite}{}}}
\subsection{Method \code{delete_favorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$delete_favorite(userid, dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-toggle_favorite"></a>}}
\if{latex}{\out{\hypertarget{method-NA-toggle_favorite}{}}}
\subsection{Method \code{toggle_favorite()}}{
Toggle favorie dossier for a user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$toggle_favorite(userid, dossierid)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Delete the favorite if the dossier is a favorite, otherwise add it.
}

\subsection{Returns}{
TRUE if successful.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-next_dossier_id"></a>}}
\if{latex}{\out{\hypertarget{method-NA-next_dossier_id}{}}}
\subsection{Method \code{next_dossier_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$next_dossier_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_dossier}{}}}
\subsection{Method \code{add_dossier()}}{
Make a new dossier, write to DB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$add_dossier(
  dossiername,
  userid,
  aanmaakdatum,
  verloopdatum,
  betrokkeninstanties,
  aanmelder,
  id = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dossiername}}{Name of the dossier (char)}

\item{\code{userid}}{Current user (char)}

\item{\code{aanmaakdatum}}{Creation date (posixct)}

\item{\code{verloopdatum}}{Expiry date (posixct)}

\item{\code{betrokkeninstanties}}{(char)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The (usually new) ID of the dossier.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_dossier}{}}}
\subsection{Method \code{get_dossier()}}{
Get a dossier based on its ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_dossier(dossierid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dossierid}}{Dossier ID's (can be a vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_my_active_dossiers"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_my_active_dossiers}{}}}
\subsection{Method \code{get_my_active_dossiers()}}{
Retrieves dossiers for a user 'Actief' or 'Volgen' status for
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_my_active_dossiers(userid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_all_active_dossiers"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_all_active_dossiers}{}}}
\subsection{Method \code{get_all_active_dossiers()}}{
Find all dossiers with 'Actief' or 'Volgen' status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_all_active_dossiers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_verwijderde_dossiers"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_verwijderde_dossiers}{}}}
\subsection{Method \code{get_verwijderde_dossiers()}}{
Find all dossiers with 'Verwijderd' status
Retrieve all pseudo-BSNs of all persons in active dossiers for the current user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_verwijderde_dossiers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_personen_active_dossiers"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_personen_active_dossiers}{}}}
\subsection{Method \code{get_personen_active_dossiers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_personen_active_dossiers(
  userid = NULL,
  method = c("current_user", "all")
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of pseudo-BSNs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_objecten_objectinstance"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_objecten_objectinstance}{}}}
\subsection{Method \code{get_objecten_objectinstance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_objecten_objectinstance(ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_dossierid_personen"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_dossierid_personen}{}}}
\subsection{Method \code{get_dossierid_personen()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_dossierid_personen(pseudoid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-dossier_naam_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-NA-dossier_naam_from_id}{}}}
\subsection{Method \code{dossier_naam_from_id()}}{
Find the name of a dossier based on its ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$dossier_naam_from_id(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-delete_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-NA-delete_dossier}{}}}
\subsection{Method \code{delete_dossier()}}{
Really delete a dossier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$delete_dossier(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-verwijder_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-NA-verwijder_dossier}{}}}
\subsection{Method \code{verwijder_dossier()}}{
Set status dossier to 'Verwijderd'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$verwijder_dossier(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-undo_verwijder_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-NA-undo_verwijder_dossier}{}}}
\subsection{Method \code{undo_verwijder_dossier()}}{
Undo verwijder_dossier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$undo_verwijder_dossier(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-dossier_exists"></a>}}
\if{latex}{\out{\hypertarget{method-NA-dossier_exists}{}}}
\subsection{Method \code{dossier_exists()}}{
Does the given Dossier ID actually exist?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$dossier_exists(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_instantie"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_instantie}{}}}
\subsection{Method \code{add_instantie()}}{
Betrokken instantie toevoegen
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$add_instantie(instantie, properties = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-list_instanties"></a>}}
\if{latex}{\out{\hypertarget{method-NA-list_instanties}{}}}
\subsection{Method \code{list_instanties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$list_instanties()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-log_audit"></a>}}
\if{latex}{\out{\hypertarget{method-NA-log_audit}{}}}
\subsection{Method \code{log_audit()}}{
Log an audit action
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$log_audit(userid, dossierid, actie, pseudo_bsn = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_audit"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_audit}{}}}
\subsection{Method \code{get_audit()}}{
Get dossier kenmerken for a dossierid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_audit(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-is_object_in_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-NA-is_object_in_dossier}{}}}
\subsection{Method \code{is_object_in_dossier()}}{
Is this objectinstance a member of dossier?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$is_object_in_dossier(objectinstance, dossierid)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if object is in the dossier, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_object_to_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_object_to_dossier}{}}}
\subsection{Method \code{add_object_to_dossier()}}{
Add person(s) (who has a BSN) to a dossier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$add_object_to_dossier(
  ids,
  dossierid,
  userid,
  type = c("bsn_persoon", "onbekend_persoon", "bag_locatie", "onbekend_locatie"),
  dossierid_check = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{Vector of pseudo-BSNs, or persoon-IDs (in case of unknown persons)}

\item{\code{dossierid}}{Dossier ID (char)}

\item{\code{userid}}{The current user (for audit trail)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if successful, FALSE if object already added to dossier.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-remove_object_from_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-NA-remove_object_from_dossier}{}}}
\subsection{Method \code{remove_object_from_dossier()}}{
Remove an object from a dossier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$remove_object_from_dossier(ids, dossierid, userid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_persoon_kenmerken"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_persoon_kenmerken}{}}}
\subsection{Method \code{add_persoon_kenmerken()}}{
Add kenmerken to a persoon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$add_persoon_kenmerken(
  dossierid,
  id,
  geboortedatum = "",
  geslacht = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_unknown_person"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_unknown_person}{}}}
\subsection{Method \code{add_unknown_person()}}{
Register a new 'unknown' person
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$add_unknown_person(
  name,
  adresseerbaarobject = NULL,
  locatieid = "",
  woonadresomschrijving = ""
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The new ID of the person
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_persoon"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_persoon}{}}}
\subsection{Method \code{get_persoon()}}{
Get a person (an unknown person) by ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_persoon(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_unknown_persoon"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_unknown_persoon}{}}}
\subsection{Method \code{edit_unknown_persoon()}}{
Update name, BSN of an unknown person
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$edit_unknown_persoon(id, naam, bsn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_unknown_location"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_unknown_location}{}}}
\subsection{Method \code{add_unknown_location()}}{
Register an 'unknown' location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$add_unknown_location(
  locatieomschrijving,
  wijk = "",
  buurt = "",
  postcode = "",
  straatnaam = "",
  huisnummer = "",
  huisnummertoevoeging = "",
  plaatsnaam = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{locatieomschrijving}}{Description of the location - mandatory}

\item{\code{wijk}}{Optional}

\item{\code{buurt}}{Optional}

\item{\code{postcode}}{Optional}

\item{\code{straatnaam}}{Optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new ID of the location
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-delete_unknown_location"></a>}}
\if{latex}{\out{\hypertarget{method-NA-delete_unknown_location}{}}}
\subsection{Method \code{delete_unknown_location()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$delete_unknown_location(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_unknown_location"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_unknown_location}{}}}
\subsection{Method \code{get_unknown_location()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_unknown_location(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_adres_as_hoofdadres"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_adres_as_hoofdadres}{}}}
\subsection{Method \code{set_adres_as_hoofdadres()}}{
Make an adres the hoofdadres of the dossier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$set_adres_as_hoofdadres(dossierid, objectinstance)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_object_relations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_object_relations}{}}}
\subsection{Method \code{set_object_relations()}}{
Fill 'relation' and 'related_to' column for an object (typically, person) 
in a dossier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$set_object_relations(
  dossierid,
  objectinstance,
  relation,
  related_to
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_object_kenmerk"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_object_kenmerk}{}}}
\subsection{Method \code{get_object_kenmerk()}}{
Get kenmerk column for an object in a dossier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_object_kenmerk(dossierid, objectinstance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dossierid}}{The dossier ID (integer)}

\item{\code{objectinstance}}{The object ID instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_object_kenmerk"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_object_kenmerk}{}}}
\subsection{Method \code{set_object_kenmerk()}}{
Fill 'kenmerk' column for an object in a dossier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$set_object_kenmerk(dossierid, objectinstance, kenmerk)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{kenmerk}}{Text field (no validation)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_objects"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_objects}{}}}
\subsection{Method \code{get_objects()}}{
Get all objects in a dossier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_objects(dossierid, type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dossierid}}{The dossier ID (integer)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-n_personen_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-NA-n_personen_dossier}{}}}
\subsection{Method \code{n_personen_dossier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$n_personen_dossier(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_dossier_kenmerken"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_dossier_kenmerken}{}}}
\subsection{Method \code{add_dossier_kenmerken()}}{
Save kenmerktypes for a dossier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$add_dossier_kenmerken(dossierid, kenmerktypeid, userid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dossierid}}{The dossier ID}

\item{\code{A}}{vector of kenmerktypes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_dossier_kenmerken"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_dossier_kenmerken}{}}}
\subsection{Method \code{get_dossier_kenmerken()}}{
Get dossier kenmerken for a dossierid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_dossier_kenmerken(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-replace_dossier_kenmerken"></a>}}
\if{latex}{\out{\hypertarget{method-NA-replace_dossier_kenmerken}{}}}
\subsection{Method \code{replace_dossier_kenmerken()}}{
Replace kenmerken for a dossier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$replace_dossier_kenmerken(dossierid, kenmerktypeid, userid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_dossierid_dossier_kenmerken"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_dossierid_dossier_kenmerken}{}}}
\subsection{Method \code{get_dossierid_dossier_kenmerken()}}{
Get dossierid's for given dossierkenmerken
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_dossierid_dossier_kenmerken(kenmerken)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coordinatoren}}{A vector of coordinatoren userid's.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of dossierid's
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_coordinatoren"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_coordinatoren}{}}}
\subsection{Method \code{add_coordinatoren()}}{
Write coordinatoren for a dossier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$add_coordinatoren(dossierid, userid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-delete_dossier_coordinatoren"></a>}}
\if{latex}{\out{\hypertarget{method-NA-delete_dossier_coordinatoren}{}}}
\subsection{Method \code{delete_dossier_coordinatoren()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$delete_dossier_coordinatoren(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-replace_dossier_coordinatoren"></a>}}
\if{latex}{\out{\hypertarget{method-NA-replace_dossier_coordinatoren}{}}}
\subsection{Method \code{replace_dossier_coordinatoren()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$replace_dossier_coordinatoren(dossierid, userid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_coordinatoren"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_coordinatoren}{}}}
\subsection{Method \code{get_coordinatoren()}}{
Get coordinatoren for a dossierid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_coordinatoren(dossierid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_dossierid_coordinatoren"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_dossierid_coordinatoren}{}}}
\subsection{Method \code{get_dossierid_coordinatoren()}}{
Get dossierid's for given coordinatoren
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$get_dossierid_coordinatoren(coordinatoren)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coordinatoren}}{A vector of coordinatoren userid's.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of dossierid's
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-encrypt"></a>}}
\if{latex}{\out{\hypertarget{method-NA-encrypt}{}}}
\subsection{Method \code{encrypt()}}{
Encrypt a value using a secret
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$encrypt(x, secret = getOption("linkit_secret"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of values to encrypt}

\item{\code{secret}}{The secret to use, set options(linkit_secret = "abc")}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A vectorized version of `safer::encrypt_string`
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-decrypt"></a>}}
\if{latex}{\out{\hypertarget{method-NA-decrypt}{}}}
\subsection{Method \code{decrypt()}}{
Decrypt a value using a secret
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$decrypt(x, secret = getOption("linkit_secret"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of values to decrypt}

\item{\code{secret}}{The secret to use, set options(linkit_secret = "abc")}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A vectorized version of `safer::decrypt_string`.
Missing values and empty strings are returned unchanged.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkItEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
