% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkitIZMNotifier.R
\name{LinkitIZMNotifier}
\alias{LinkitIZMNotifier}
\title{Implementation of a notification system between IZM and LinkIt}
\description{
Implementation of a notification system between IZM and LinkIt

Implementation of a notification system between IZM and LinkIt
}
\section{Super class}{
\code{\link[linkitr:NA]{linkitr::NA}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LinkitIZMNotifier$new()}}
\item \href{#method-get_notificaties}{\code{LinkitIZMNotifier$get_notificaties()}}
\item \href{#method-remove_notificatie}{\code{LinkitIZMNotifier$remove_notificatie()}}
\item \href{#method-person_is_in_linkit_dossier}{\code{LinkitIZMNotifier$person_is_in_linkit_dossier()}}
\item \href{#method-person_has_recent_notificatie}{\code{LinkitIZMNotifier$person_has_recent_notificatie()}}
\item \href{#method-append_notificatie}{\code{LinkitIZMNotifier$append_notificatie()}}
\item \href{#method-clone}{\code{LinkitIZMNotifier$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="add_coordinatoren">}\href{../../linkitr/html/NA.html#method-add_coordinatoren}{\code{linkitr::NA$add_coordinatoren()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="add_dossier">}\href{../../linkitr/html/NA.html#method-add_dossier}{\code{linkitr::NA$add_dossier()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="add_dossier_kenmerken">}\href{../../linkitr/html/NA.html#method-add_dossier_kenmerken}{\code{linkitr::NA$add_dossier_kenmerken()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="add_favorite">}\href{../../linkitr/html/NA.html#method-add_favorite}{\code{linkitr::NA$add_favorite()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="add_object_to_dossier">}\href{../../linkitr/html/NA.html#method-add_object_to_dossier}{\code{linkitr::NA$add_object_to_dossier()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="add_persoon_kenmerken">}\href{../../linkitr/html/NA.html#method-add_persoon_kenmerken}{\code{linkitr::NA$add_persoon_kenmerken()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="add_unknown_location">}\href{../../linkitr/html/NA.html#method-add_unknown_location}{\code{linkitr::NA$add_unknown_location()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="add_unknown_person">}\href{../../linkitr/html/NA.html#method-add_unknown_person}{\code{linkitr::NA$add_unknown_person()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="add_user">}\href{../../linkitr/html/NA.html#method-add_user}{\code{linkitr::NA$add_user()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="append_data">}\href{../../linkitr/html/NA.html#method-append_data}{\code{linkitr::NA$append_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="close">}\href{../../linkitr/html/NA.html#method-close}{\code{linkitr::NA$close()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="decrypt">}\href{../../linkitr/html/NA.html#method-decrypt}{\code{linkitr::NA$decrypt()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="delete_dossier">}\href{../../linkitr/html/NA.html#method-delete_dossier}{\code{linkitr::NA$delete_dossier()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="delete_dossier_coordinatoren">}\href{../../linkitr/html/NA.html#method-delete_dossier_coordinatoren}{\code{linkitr::NA$delete_dossier_coordinatoren()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="delete_favorite">}\href{../../linkitr/html/NA.html#method-delete_favorite}{\code{linkitr::NA$delete_favorite()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="delete_rows_where">}\href{../../linkitr/html/NA.html#method-delete_rows_where}{\code{linkitr::NA$delete_rows_where()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="delete_rows_where2">}\href{../../linkitr/html/NA.html#method-delete_rows_where2}{\code{linkitr::NA$delete_rows_where2()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="delete_unknown_location">}\href{../../linkitr/html/NA.html#method-delete_unknown_location}{\code{linkitr::NA$delete_unknown_location()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="dossier_exists">}\href{../../linkitr/html/NA.html#method-dossier_exists}{\code{linkitr::NA$dossier_exists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="dossier_naam_from_id">}\href{../../linkitr/html/NA.html#method-dossier_naam_from_id}{\code{linkitr::NA$dossier_naam_from_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="edit_unknown_persoon">}\href{../../linkitr/html/NA.html#method-edit_unknown_persoon}{\code{linkitr::NA$edit_unknown_persoon()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="encrypt">}\href{../../linkitr/html/NA.html#method-encrypt}{\code{linkitr::NA$encrypt()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="execute_query">}\href{../../linkitr/html/NA.html#method-execute_query}{\code{linkitr::NA$execute_query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="format_dossierid">}\href{../../linkitr/html/NA.html#method-format_dossierid}{\code{linkitr::NA$format_dossierid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_all_active_dossiers">}\href{../../linkitr/html/NA.html#method-get_all_active_dossiers}{\code{linkitr::NA$get_all_active_dossiers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_audit">}\href{../../linkitr/html/NA.html#method-get_audit}{\code{linkitr::NA$get_audit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_con">}\href{../../linkitr/html/NA.html#method-get_con}{\code{linkitr::NA$get_con()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_coordinatoren">}\href{../../linkitr/html/NA.html#method-get_coordinatoren}{\code{linkitr::NA$get_coordinatoren()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_dossier">}\href{../../linkitr/html/NA.html#method-get_dossier}{\code{linkitr::NA$get_dossier()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_dossier_kenmerken">}\href{../../linkitr/html/NA.html#method-get_dossier_kenmerken}{\code{linkitr::NA$get_dossier_kenmerken()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_dossierid_coordinatoren">}\href{../../linkitr/html/NA.html#method-get_dossierid_coordinatoren}{\code{linkitr::NA$get_dossierid_coordinatoren()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_dossierid_dossier_kenmerken">}\href{../../linkitr/html/NA.html#method-get_dossierid_dossier_kenmerken}{\code{linkitr::NA$get_dossierid_dossier_kenmerken()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_favorites">}\href{../../linkitr/html/NA.html#method-get_favorites}{\code{linkitr::NA$get_favorites()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_my_active_dossiers">}\href{../../linkitr/html/NA.html#method-get_my_active_dossiers}{\code{linkitr::NA$get_my_active_dossiers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_object_kenmerk">}\href{../../linkitr/html/NA.html#method-get_object_kenmerk}{\code{linkitr::NA$get_object_kenmerk()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_objects">}\href{../../linkitr/html/NA.html#method-get_objects}{\code{linkitr::NA$get_objects()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_personen_active_dossiers">}\href{../../linkitr/html/NA.html#method-get_personen_active_dossiers}{\code{linkitr::NA$get_personen_active_dossiers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_persoon">}\href{../../linkitr/html/NA.html#method-get_persoon}{\code{linkitr::NA$get_persoon()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_unknown_location">}\href{../../linkitr/html/NA.html#method-get_unknown_location}{\code{linkitr::NA$get_unknown_location()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="get_verwijderde_dossiers">}\href{../../linkitr/html/NA.html#method-get_verwijderde_dossiers}{\code{linkitr::NA$get_verwijderde_dossiers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="has_value">}\href{../../linkitr/html/NA.html#method-has_value}{\code{linkitr::NA$has_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="is_admin">}\href{../../linkitr/html/NA.html#method-is_admin}{\code{linkitr::NA$is_admin()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="is_favorite">}\href{../../linkitr/html/NA.html#method-is_favorite}{\code{linkitr::NA$is_favorite()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="is_object_in_dossier">}\href{../../linkitr/html/NA.html#method-is_object_in_dossier}{\code{linkitr::NA$is_object_in_dossier()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="laatste_login">}\href{../../linkitr/html/NA.html#method-laatste_login}{\code{linkitr::NA$laatste_login()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="list_tables">}\href{../../linkitr/html/NA.html#method-list_tables}{\code{linkitr::NA$list_tables()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="log_audit">}\href{../../linkitr/html/NA.html#method-log_audit}{\code{linkitr::NA$log_audit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="make_choices">}\href{../../linkitr/html/NA.html#method-make_choices}{\code{linkitr::NA$make_choices()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="next_dossier_id">}\href{../../linkitr/html/NA.html#method-next_dossier_id}{\code{linkitr::NA$next_dossier_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="query">}\href{../../linkitr/html/NA.html#method-query}{\code{linkitr::NA$query()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="read_table">}\href{../../linkitr/html/NA.html#method-read_table}{\code{linkitr::NA$read_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="remove_object_from_dossier">}\href{../../linkitr/html/NA.html#method-remove_object_from_dossier}{\code{linkitr::NA$remove_object_from_dossier()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="replace_dossier_coordinatoren">}\href{../../linkitr/html/NA.html#method-replace_dossier_coordinatoren}{\code{linkitr::NA$replace_dossier_coordinatoren()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="replace_dossier_kenmerken">}\href{../../linkitr/html/NA.html#method-replace_dossier_kenmerken}{\code{linkitr::NA$replace_dossier_kenmerken()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="replace_value_where">}\href{../../linkitr/html/NA.html#method-replace_value_where}{\code{linkitr::NA$replace_value_where()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="set_adres_as_hoofdadres">}\href{../../linkitr/html/NA.html#method-set_adres_as_hoofdadres}{\code{linkitr::NA$set_adres_as_hoofdadres()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="set_object_kenmerk">}\href{../../linkitr/html/NA.html#method-set_object_kenmerk}{\code{linkitr::NA$set_object_kenmerk()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="set_object_relations">}\href{../../linkitr/html/NA.html#method-set_object_relations}{\code{linkitr::NA$set_object_relations()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="toggle_favorite">}\href{../../linkitr/html/NA.html#method-toggle_favorite}{\code{linkitr::NA$toggle_favorite()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="undo_verwijder_dossier">}\href{../../linkitr/html/NA.html#method-undo_verwijder_dossier}{\code{linkitr::NA$undo_verwijder_dossier()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="username_from_userid">}\href{../../linkitr/html/NA.html#method-username_from_userid}{\code{linkitr::NA$username_from_userid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="linkitr" data-topic="NA" data-id="verwijder_dossier">}\href{../../linkitr/html/NA.html#method-verwijder_dossier}{\code{linkitr::NA$verwijder_dossier()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkitIZMNotifier$new(
  config_name,
  schema,
  pool,
  config_file = "conf/config.yml"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_notificaties"></a>}}
\if{latex}{\out{\hypertarget{method-get_notificaties}{}}}
\subsection{Method \code{get_notificaties()}}{
Retrieve IZM notifications for all dossiers of which the 
current user is the coordinator, and the dossier is active,
and the notification has not been removed yet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkitIZMNotifier$get_notificaties(userid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{Make a notification invisible}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_notificatie"></a>}}
\if{latex}{\out{\hypertarget{method-remove_notificatie}{}}}
\subsection{Method \code{remove_notificatie()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkitIZMNotifier$remove_notificatie(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id of the notificatie}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-person_is_in_linkit_dossier"></a>}}
\if{latex}{\out{\hypertarget{method-person_is_in_linkit_dossier}{}}}
\subsection{Method \code{person_is_in_linkit_dossier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkitIZMNotifier$person_is_in_linkit_dossier(pseudo_bsn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-person_has_recent_notificatie"></a>}}
\if{latex}{\out{\hypertarget{method-person_has_recent_notificatie}{}}}
\subsection{Method \code{person_has_recent_notificatie()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkitIZMNotifier$person_has_recent_notificatie(pseudo_bsn, recent_days = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_notificatie"></a>}}
\if{latex}{\out{\hypertarget{method-append_notificatie}{}}}
\subsection{Method \code{append_notificatie()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkitIZMNotifier$append_notificatie(pseudo_bsn, userid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinkitIZMNotifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
